nice

+++ ft_print_hex.c
@@ -6,17 +6,17 @@
 /*   By: houmanso <houmanso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 16:48:30 by houmanso          #+#    #+#             */
-/*   Updated: 2022/11/08 18:16:31 by houmanso         ###   ########.fr       */
+/*   Updated: 2022/11/08 18:51:00 by houmanso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_print_hex(unsigned int n, char *base ,size_t *len)
+void	ft_print_hex(size_t n, char *base ,size_t *len)
 {
-	unsigned int	nbr;
+	size_t	nbr;
 
-	nbr = (unsigned int) n;
+	nbr = n;
 	if (nbr >= 16)
 	{
 		ft_print_hex(nbr / 16, base, len);
